requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.1
cssbeautifier==1.14.7
Django==4.1.7
django-ace==1.15.3
django-htmx==1.14.0
django-taggit==3.1.0
djlint==1.19.16
EditorConfig==0.12.3
factory-boy==3.2.1
Faker==17.6.0
flake8==6.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
inflection==0.5.1
iniconfig==2.0.0
jsbeautifier==1.14.7
Markdown==3.4.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
regex==2022.10.31
six==1.16.0
sqlparse==0.4.3
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.7
